+++ queue/app.go
@@ -9,7 +9,7 @@ type App interface {
 	GetInstanceId() string
 	IsReady() bool
 	GetOutboundResourceById(id int64) (*model.OutboundResource, *model.AppError)
-	GetQueueById(id int) (*model.Queue, *model.AppError)
+	GetQueueById(id int64) (*model.Queue, *model.AppError)
 	NewCall(params *model.CallRequest) (uuid, cause string, err *model.AppError)
 
 	ConsumeCallEvent() <-chan mq.Event
